requirements.txt
rasa==3.6.15
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pdfplumber==0.9.0
psycopg2-binary==2.9.9
tensorflow==2.12.0 \
  --hash=sha256:e1f73d248894c9d77766c4079a0705bd5e781523487e0f1be30f3cd90fb4a1b6
